requirements.txt
# Core dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# API clients
tweepy==4.14.0
anthropic==0.7.8
httpx==0.25.2

# Data handling
pydantic==2.5.0
pyyaml==6.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Async support
asyncio-throttle==1.0.2
aiofiles==23.2.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
typer==0.9.0

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1